# static/api/openapi.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "Tyman.tech API",
		"description": "The api for the tyman.tech website",
		"version": "1.0.3"
	},
	"servers": [
		{
			"url": "https://tyman.tech/api",
			"description": "Main server"
		},
		{
			"url": "http://localhost:{port}/api",
			"description": "Local dev server",
			"variables": {
				"port": {
					"default": "8738",
					"description": "Local test server port"
				}
			}
		}
	],
	"tags": [
		{
			"name": "Images",
			"description": "API Endpoints for images"
		},
		{
			"name": "Mods",
			"description": "API Endpoints for minecraft mods"
		},
		{
			"name": "Links",
			"description": "API Endpoints for linked accounts"
		}
	],
	"paths": {
		"/links/minecraft/{discordID}": {
			"get": {
				"description": "Get a minecraft account from discord ID",
				"operationId": "getMinecraftFromDiscord",
				"parameters": [
					{
						"name": "discordID",
						"in": "path",
						"required": true,
						"style": "simple",
						"explode": false,
						"schema": {
							"type": "string",
							"description": "The discord ID of the user to fetch"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json; charset=utf-8": {
								"schema": {
									"$ref": "#/components/schemas/Discord To Minecraft"
								}
							}
						},
						"description": "Success"
					},
					"400": {
						"content": {
							"application/json; charset=utf-8": {
								"schema": {
									"type": "object",
									"required": ["success", "reason"],
									"properties": {
										"success": {
											"type": "boolean",
											"description": "Whether or not the request succeeded or not",
											"example": true
										},
										"reason": {
											"type": "string",
											"description": "The reason of failure",
											"example": "Link not found"
										}
									}
								}
							}
						},
						"description": "Not found"
					}
				},
				"tags": ["Links"]
			}
		},
		"/links/minecraft": {
			"post": {
				"description": "Add a mc and discord account to db (requires admin key because I'm too lazy to add oauth support right now)",
				"operationId": "setMinecraftAndDiscord",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": ["discordID", "minecraftID"],
								"properties": {
									"discordID": {
										"type": "string",
										"description": "The discord ID of the user",
										"example": "487443883127472129"
									},
									"minecraftID": {
										"type": "string",
										"description": "The minecraft UUID of the user",
										"example": "424ef6d0-4774-4f8c-8bef-8f62ebdac9c0"
									}
								}
							}
						}
					}
				},
				"security": [
					{
						"API Key": []
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json; charset=utf-8": {
								"schema": {
									"type": "object",
									"required": ["success"],
									"properties": {
										"success": {
											"type": "boolean",
											"description": "Whether or not the request succeeded or not",
											"example": true
										}
									}
								}
							}
						},
						"description": "Success"
					}
				},
				"tags": ["Links"]
			}
		},
		"/files/mods": {
			"get": {
				"tags": ["Mods"],
				"summary": "Get mods",
				"description": "Gets all the availible mod files",
				"operationId": "getMods",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json; charset=utf-8": {
								"schema": {
									"$ref": "#/components/schemas/Mod Files"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json; charset=utf-8": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"502": {
						"description": "Bad Gateway",
						"content": {
							"text/plain; charset=utf-8": {
								"schema": {
									"$ref": "#/components/schemas/Bad Gateway"
								}
							}
						}
					}
				}
			}
		},
		"/files/images": {
			"get": {
				"summary": "Get images",
				"description": "Get a list of images that have been uplaoded",
				"operationId": "getImages",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json; charset=utf-8": {
								"schema": {
									"$ref": "#/components/schemas/Images List"
								}
							}
						}
					},
					"403": {
						"description": "Unauthorized",
						"content": {
							"application/json; charset=utf-8": {
								"schema": {
									"$ref": "#/components/schemas/Unauthorized"
								}
							}
						}
					},
					"502": {
						"description": "Bad Gateway",
						"content": {
							"text/plain; charset=utf-8": {
								"schema": {
									"$ref": "#/components/schemas/Bad Gateway"
								}
							}
						}
					}
				},
				"security": [
					{
						"API Key": []
					}
				],
				"tags": ["Images"]
			},
			"post": {
				"tags": ["Images"],
				"summary": "Upload image",
				"description": "Upload an image to the site",
				"operationId": "uploadImage",
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json; charset=utf-8": {
								"schema": {
									"$ref": "#/components/schemas/File Upload"
								}
							}
						}
					},
					"403": {
						"description": "Unauthorized",
						"content": {
							"application/json; charset=utf-8": {
								"schema": {
									"$ref": "#/components/schemas/Unauthorized"
								}
							}
						}
					},
					"422": {
						"description": "File not given",
						"content": {
							"application/json; charset=utf-8": {
								"schema": {
									"$ref": "#/components/schemas/File Upload 422"
								}
							}
						}
					},
					"502": {
						"description": "Bad Gateway",
						"content": {
							"text/plain; charset=utf-8": {
								"schema": {
									"$ref": "#/components/schemas/Bad Gateway"
								}
							}
						}
					}
				},
				"security": [
					{
						"API Key": []
					}
				]
			}
		},
		"/files/images/{file}": {
			"delete": {
				"tags": ["Images"],
				"summary": "Delete image",
				"description": "Delete an image file",
				"operationId": "deleteImage",
				"parameters": [
					{
						"name": "file",
						"in": "path",
						"required": true,
						"style": "simple",
						"explode": false,
						"schema": {
							"type": "string",
							"description": "The filename of the file to delete"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json; charset=utf-8": {
								"schema": {
									"type": "object",
									"properties": {
										"success": {
											"type": "boolean",
											"description": "Whether or not the request succeeded or not",
											"example": true
										}
									}
								}
							}
						}
					},
					"403": {
						"description": "Unauthorized",
						"content": {
							"application/json; charset=utf-8": {
								"schema": {
									"$ref": "#/components/schemas/Unauthorized"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json; charset=utf-8": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"502": {
						"description": "Bad Gateway",
						"content": {
							"text/plain; charset=utf-8": {
								"schema": {
									"$ref": "#/components/schemas/Bad Gateway"
								}
							}
						}
					}
				},
				"security": [
					{
						"API Key": []
					}
				]
			}
		}
	},
	"components": {
		"schemas": {
			"Discord To Minecraft": {
				"type": "object",
				"required": ["success", "discordID", "minecraftID"],
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Whether or not the request succeeded or not",
						"example": true
					},
					"discordID": {
						"type": "string",
						"description": "The discord id of the user",
						"example": "487443883127472129"
					},
					"minecraftID": {
						"type": "string",
						"description": "The minecraft uuid of the user (with dashes)",
						"example": "424ef6d0-4774-4f8c-8bef-8f62ebdac9c0"
					}
				}
			},
			"File Upload": {
				"type": "object",
				"required": ["success", "url", "id", "filename"],
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Whether or not the request succeeded or not",
						"example": true
					},
					"url": {
						"type": "string",
						"description": "The url of the uploaded image",
						"example": "https://tyman.tech/394nRWnVRe.png"
					},
					"filename": {
						"type": "string",
						"description": "The filename on the server",
						"example": "394nRWnVRe.png"
					},
					"id": {
						"type": "string",
						"description": "The ID of the file on the server",
						"example": "394nRWnVRe"
					}
				}
			},
			"File Upload 422": {
				"type": "object",
				"required": ["success", "reason"],
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Whether or not the request succeeded or not",
						"example": false
					},
					"reason": {
						"type": "string",
						"description": "The reason of failure",
						"example": "File not given"
					}
				}
			},
			"Mod Files": {
				"type": "object",
				"required": ["success", "files"],
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Whether or not the request succeeded or not",
						"example": true
					},
					"files": {
						"type": "array",
						"description": "The array of file objects",
						"items": {
							"type": "object",
							"required": [
								"name",
								"version",
								"description",
								"needs-whitelist",
								"url"
							],
							"properties": {
								"name": {
									"type": "string",
									"description": "The name of the mod",
									"example": "Example Mod"
								},
								"version": {
									"type": "string",
									"description": "The version of the mod",
									"example": "1.6.9"
								},
								"semver": {
									"type": "string",
									"description": "The semver-compatable version of the mod (only if the `version` property itself is not semver-compatible)",
									"example": "1.6.9"
								},
								"description": {
									"type": "string",
									"description": "The description of the mod",
									"example": "An example mod"
								},
								"needs-whitelist": {
									"type": "boolean",
									"description": "Whether or not this mod needs the user to be whitelisted to be used",
									"example": false
								},
								"url": {
									"type": "string",
									"description": "The url containing the file",
									"example": "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
								}
							}
						}
					}
				}
			},
			"Error": {
				"type": "object",
				"required": ["success", "reason"],
				"description": "Internal Server Error",
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Whether or not the request succeeded or not",
						"example": false
					},
					"reason": {
						"type": "string",
						"description": "The reason of failure",
						"example": "Internal Server Error"
					}
				}
			},
			"Unauthorized": {
				"type": "object",
				"description": "Unauthorized",
				"required": ["success", "reason"],
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Whether or not the request succeeded or not",
						"example": false
					},
					"reason": {
						"type": "string",
						"description": "The reason of failure",
						"example": "Invalid key"
					}
				}
			},
			"Bad Gateway": {
				"type": "string",
				"description": "Bad Gateway",
				"example": "<insert cloudflare html page here>"
			},
			"Images List": {
				"type": "object",
				"required": ["success", "files"],
				"properties": {
					"success": {
						"type": "boolean",
						"description": "Whether or not the request succeeded or not",
						"example": true
					},
					"files": {
						"type": "array",
						"description": "The file list",
						"items": {
							"type": "string",
							"description": "The name of the file",
							"example": "Vl6ea8rHHB.png"
						}
					}
				}
			}
		},
		"securitySchemes": {
			"API Key": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	}
}